requirements.txt
Django==3.0.4
django-modeltranslation==0.14.4
django-audiofield==0.10.0
